core/lib/settings: Generalized Settings Manager to work on any platform
This commit moves the Settings Manager from the AVR codebase
into the Contiki core library. Any platform that implements
the Contiki EEPROM API can now use the Settings Manager's
key-value store for storing their persistent configuration info.

The Settings Manager is a EEPROM-based key-value store. Keys
are 16-bit integers and values may be up to 16,383 bytes long.
It is intended to be used to store configuration-related information,
like network settings, radio channels, etc.

 * Robust data format which requires no initialization.
 * Supports multiple values with the same key.
 * Data can be appended without erasing EEPROM.
 * Max size of settings data can be easily increased in the future,
   as long as it doesn't overlap with application data.

The format was inspired by the [OLPC manufacturing data format][].

Since the beginning of EEPROM often contains application-specific
information, the best place to store settings is at the end of
EEPROM (the "top"). Because we are starting at the end of EEPROM, it makes sense
to grow the list of key-value pairs downward, toward the start of
EEPROM.

Each key-value pair is stored in memory in the following format:

<table>
  <thead>
   <td>Order</td>
   <td>Size<small> (in bytes)</small></td>
   <td>Name</td>
   <td>Description</td>
  </thead>
  <tr>
   <td>0</td>
   <td>2</td>
   <td><code>key</code></td>
   <td>16-bit key</td>
  </tr>
  <tr>
   <td>-2</td>
   <td>1</td>
   <td><code>size_check</code></td>
   <td>One's-complement of next byte</td>
  </tr>
  <tr>
   <td>-3</td>
   <td>1 or 2</td>
   <td><code>size</code></td>
   <td>The size of the value, in bytes</td>
  </tr>
  <tr>
   <td>-4 or -5</td>
   <td>variable</td>
   <td><code>value</code></td>
   <td>Value associated with key</td>
  </tr>
</table>

The end of the key-value pairs is denoted by the first invalid entry.
An invalid entry has any of the following attributes:

 * The size_check byte doesn't match the one's compliment
   of the size byte (or `size_low` byte).
 * The key has a value of 0x0000.

[OLPC manufacturing data format]: http://wiki.laptop.org/go/Manufacturing_data