jackdaw: Move cdc_task and sicslow_ethernet over to ravenusbstick.
The "Jackdaw" firmware for the Atmel RZUSBStick could be very useful
on other platforms, but it's implementation is highly integrated into
the `avr-ravenusb` platform. This commit takes the first step toward
changing that, by moving the cdc_task and sicslow_ethernet code out
of the `avr-ravenusb` platform and into the `ravenusbstick` example.